+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Unreleased
+## 2025.9.0
 
 - Added "Rebuild Ctags" editor Command. The old method of building ctags via a Task is going to be deprecated.